ADD: rendering of pacman image in accordance with movement direction
The goal is to have the Pacman object to require as less information about the rendering process as possible.
It is simple a state object. It does not own it's texture since it is owned by the renderer due to efficient
hardware rendering.
SpriteSheet is not needed in this new design